Extract `SensorSetter` from NativeProxy.h
